Separating component into multiple files. updating search bar action.
